+++ src/java/abfab3d/grid/op/DilationCube.java
@@ -24,7 +24,7 @@ import abfab3d.grid.*;
  * @author Tony Wong
  */
 public class DilationCube implements Operation, AttributeOperation {
-	
+
     /** The distance from a voxel to dilate */
     private int distance;
 
@@ -40,12 +40,12 @@ public class DilationCube implements Operation, AttributeOperation {
      * @return The new grid
      */
     public Grid execute(Grid dest) {
-    	
-    	// Nothing to do if distance is 0
-    	if (distance == 0) {
-    		return dest;
-    	}
-    	
+
+        // Nothing to do if distance is 0
+        if (distance == 0) {
+            return dest;
+        }
+
         int height = dest.getHeight();
         int width = dest.getWidth();
         int depth = dest.getDepth();
@@ -54,7 +54,7 @@ public class DilationCube implements Operation, AttributeOperation {
         // the dilation distance
         Grid dilatedGrid = dest.createEmpty(width + 2 * distance,
                                             height + 2 * distance,
-                                            depth + 2 * distance,                                            
+                                            depth + 2 * distance,
                                             dest.getVoxelSize(),
                                             dest.getSliceHeight());
 
@@ -65,7 +65,7 @@ public class DilationCube implements Operation, AttributeOperation {
                     byte state = dest.getState(x, y, z);
 
                     if (state != Grid.OUTSIDE) {
-                    	dilateVoxel(dilatedGrid, x+distance, y+distance, z+distance, state);
+                        dilateVoxel(dilatedGrid, x+distance, y+distance, z+distance, state);
                     }
                 }
             }
@@ -107,7 +107,7 @@ public class DilationCube implements Operation, AttributeOperation {
                     byte state = dest.getState(x, y, z);
 
                     if (state != Grid.OUTSIDE) {
-                        int mat = dest.getAttribute(x, y, z);
+                        long mat = dest.getAttribute(x, y, z);
                         dilateVoxel(dilatedGrid, x+distance, y+distance, z+distance, state, mat);
                     }
                 }
@@ -117,21 +117,21 @@ public class DilationCube implements Operation, AttributeOperation {
         return dilatedGrid;
     }
 
-    private void dilateVoxel(AttributeGrid grid, int xPos, int yPos, int zPos, byte state, int mat) {
+    private void dilateVoxel(AttributeGrid grid, int xPos, int yPos, int zPos, byte state, long mat) {
         int xStart = xPos - distance;
         int xEnd = xPos + distance;
         int yStart = yPos - distance;
         int yEnd = yPos + distance;
         int zStart = zPos - distance;
         int zEnd = zPos + distance;
-        
-    	for (int y=yStart; y<=yEnd; y++) {
-    		for (int x=xStart; x<=xEnd; x++) {
-    			for (int z=zStart; z<=zEnd; z++) {
-    				grid.setData(x, y, z, state, mat);
-    			}
-    		}
-    	}
+
+        for (int y=yStart; y<=yEnd; y++) {
+            for (int x=xStart; x<=xEnd; x++) {
+                for (int z=zStart; z<=zEnd; z++) {
+                    grid.setData(x, y, z, state, mat);
+                }
+            }
+        }
     }
 
     private void dilateVoxel(Grid grid, int xPos, int yPos, int zPos, byte state) {
